File persistent storage implementation

+++ src/dao/other/OtherUserDao.java
@@ -30,12 +30,8 @@ import dao.IUserDao;
  * 
  */
 public class OtherUserDao implements IUserDao {
-	private String usersPath = "persistent" + File.separator + "Users";
-	private String gamesPath = "persistent" + File.separator + "Games";
-	private String commandsPath = "persistent" + File.separator + "Commands";
 	private String usersFileStr = "persistent" + File.separator + "Users"
 			+ File.separator + "users.txt";
-	private IServerFacade facade;
 	Gson g;
 
 	public OtherUserDao() {
@@ -49,11 +45,16 @@ public class OtherUserDao implements IUserDao {
 		// out.write(g.toJson(user));
 		// }
 		try {
-			Files.write(Paths.get(usersFileStr),
-					(g.toJson(user) + "\r").getBytes(),
-					StandardOpenOption.APPEND);
-			System.out.println("user: " + g.toJson(user));
-			System.out.println("Wrote user " + user.getName() + " to file.");
+			File usersFile = new File(usersFileStr);
+			if (!usersFile.exists()) {
+				Files.write(Paths.get(usersFileStr),
+						(g.toJson(user) + "\r").getBytes(),
+						StandardOpenOption.CREATE);
+			} else {
+				Files.write(Paths.get(usersFileStr),
+						(g.toJson(user) + "\r").getBytes(),
+						StandardOpenOption.APPEND);
+			}
 		} catch (IOException e) {
 			System.out.println("Failed to write user to file.");
 			e.printStackTrace();
@@ -69,7 +70,7 @@ public class OtherUserDao implements IUserDao {
 			try {
 				userScan = new Scanner(usersFile);
 				while (userScan.hasNextLine()) {
-					usersList.add(g.fromJson(userScan.nextLine(),User.class));
+					usersList.add(g.fromJson(userScan.nextLine(), User.class));
 				}
 			} catch (FileNotFoundException e) {
 				e.printStackTrace();
